requirements (1).txt
streamlit
openai
faiss-cpu
sentence-transformers
textblob
scikit-learn
numpy
pandas
pdfplumber
python-docx
PyMuPDF
